User api with jpa and jax-rs auth filter finally works

+++ src/bean/PostMessage.java
@@ -5,17 +5,17 @@ public class PostMessage {
 	public PostMessage() {
 	}
 	
-	public PostMessage(int id) {
+	public PostMessage(long id) {
 		this.id = id;
 	}
 
-	private int id;
+	private long id;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
